requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
redis>=5.0.1
asyncpg>=0.29.0
sqlalchemy>=2.0.25
alembic>=1.13.1
celery>=5.3.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
email-validator>=2.1.0
jinja2>=3.1.3
boto3>=1.34.0
